Refactor: Update import paths for login and register forms to use new Services directory structure
